Add support for Bépo keyboard layout
This layout is a Dvorak-like initiative, used mainly in France, that can
be used to type any language with more ease that QWERTY & Co. It is
already supported out of the box by most Linux/BSD distributions.

I used
http://linuxfr.org/users/jpec/journaux/de-la-conception-d-une-disposition-bepo-pour-android
as an inspiration for a variant more adapted to phones.

See http://bepo.fr/ for more information about the layout.

Change-Id: I6a5b43608d23cf99ddde4e241e1691aae3ff8098
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>